fix: make better

+++ src/net/include/pb_conn.h
@@ -41,7 +41,7 @@ class PbConn : public NetConn {
    * The Variable need by read the buf,
    * We allocate the memory when we start the server
    */
-  uint32_t header_len_{static_cast<uint32_t>(static_cast<uint32_t>(-1))};
+  uint32_t header_len_{static_cast<uint32_t>(-1)};
   char* rbuf_;
   uint32_t cur_pos_{0};
   uint32_t rbuf_len_{0};
